feat: create purchase order no invoice

+++ src/main/java/com/swp/jewelrystore/entity/GemPriceEntity.java
@@ -28,16 +28,16 @@ public class GemPriceEntity {
     private String clarity;
 
     @Column(name = "carat_weight", nullable = false)
-    private Integer caratWeight;
+    private Double caratWeight;
 
     @Column(name = "cut", nullable = false)
     private String cut;
 
     @Column(name = "buy_price", nullable = false)
-    private Integer buyPrice;
+    private Double buyPrice;
 
     @Column(name = "sell_price", nullable = false)
-    private Integer sellPrice;
+    private Double sellPrice;
 
     @Column(name = "effect_date", nullable = false)
     private Instant effectDate;
